// apps/web-svelte/src/lib/api/generated/model/entityCard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { EntityCapability } from './entityCapability';
import type { EntityGroup } from './entityGroup';

export interface EntityCard {
  id: string;
  kind: string;
  title: string;
  /** @nullable */
  parentEntityId: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  sortOrder: number | string | null;
  capabilities: EntityCapability[];
  childrenByKind: EntityGroup[];
  relationships: EntityGroup[];
}